requirements.txt
pyautogen
ag2[openai]
python-dotenv
python-docx
fpdf
PyMuPDF
python-multipart
jinja2
httpx
fastapi
uvicorn
pytesseract
Pillow
pdfplumber
tqdm
docx2pdf

# RAG System Dependencies
faiss-cpu
rank-bm25
tiktoken
numpy
scikit-learn